# http_api/openapi3.yaml
openapi: 3.0.3
info:
  title: BitTorrent-over-SCION HTTP API
  description: TODO
  version: 0.0.1
servers:
  - url: 'http://localhost:8000'
paths:
  /api/info:
    get:
      summary: Provides information about the API and its dependencies
      responses:
        '200':
          description: successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiInfo'
        '500':
          description: internal server error, see error field for details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/torrent:
    get:
      summary: Lists all torrents with their statuses
      tags:
        - torrent
      responses:
        '200':
          description: successful
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Torrent'
        '500':
          description: internal server error, see error field for details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Adds a torrent for download and/or seeding
      tags:
        - torrent
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                torrentFile:
                  type: string
                  format: binary
                peer:
                  type: string
              required:
                - torrentFile
                - peer
              example:
                torrentFile: "<file>"
                peer: "19-ffaa:1:106d,[127.0.0.1]:43000"
      responses:
        '201':
          description: created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Torrent'
        '400':
          description: invalid torrent file, see error field for details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '415':
          description: invalid content type ("multipart/form-data" wanted)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: internal server error, see error field for details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/torrent/{id}:
    get:
      summary: Get details for a torrent by ID
      tags:
        - torrent
      parameters:
        - name: id
          in: path
          description: ID of the torrent of interest
          required: true
          schema:
            type: integer
            format: uint
      responses:
        '200':
          description: successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Torrent'
        '400':
          description: invalid ID specified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: torrent with given ID not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: internal server error, see error field for details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Deletes stopped/finished torrent
      tags:
        - torrent
      parameters:
        - name: id
          in: path
          description: ID of the torrent of interest
          required: true
          schema:
            type: integer
            format: uint
        - name: deleteFiles
          in: query
          required: false
          schema:
            type: boolean
          description: Set to true if files (downloaded files, torrent file) shall be deleted from disk as well
      responses:
        '200':
          description: deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Torrent'
        '400':
          description: invalid value specified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: torrent with given ID not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: torrent is running, stop it before deletion
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: internal server error, see error field for details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: <human-readable error message>
    ApiInfo:
      type: object
      properties:
        version:
          type: string
          example: 0.0.1
    Torrent:
      type: object
      properties:
        id:
          type: integer
          format: uint
          example: 42
        name:
          type: string
          example: "sample.torrent"
        state:
          type: string
          enum: [not started yet]